refactor(tests): optimize E2E suite from 20+ min to ~2 min
Consolidate 3 heaviest test files (product-name-consistency, links,
navigation) into content-validation.spec.js using internal iteration
pattern. Reduce default browser matrix from 5 to 2 (Chromium + Mobile
Chrome) with BROWSERS env var for flexibility. Enable parallel execution
with 2 workers and reduce timeouts. Remove duplicated tests from
i18n.spec.js and branding.spec.js.

Key changes:
- 325 test cases → ~56 (internal iteration instead of one test per page)
- 5 browsers → 2 default (BROWSERS=all for full suite)
- Sequential → parallel (fullyParallel: true, workers: 2)
- Runtime: 20+ min → ~2.2 min
